refactor(server-cloudflare): merge auth middleware into auth.ts, remove auth/ folder
The auth/ folder contained a single 21-line middleware file that was
tightly coupled to the auth singleton. Consolidating eliminates the
unnecessary indirection of a folder for one file.
