+++ packages/server-cloudflare/src/auth.ts
@@ -6,6 +6,7 @@ import { bearer } from 'better-auth/plugins/bearer';
 import { jwt } from 'better-auth/plugins/jwt';
 import { oauthProvider } from '@better-auth/oauth-provider';
 import { drizzle } from 'drizzle-orm/neon-http';
+import { factory } from './env';
 
 const sql = neon(env.DATABASE_URL);
 const db = drizzle(sql);
@@ -86,3 +87,22 @@ export const auth = betterAuth({
 		delete: (key) => env.SESSION_KV.delete(key),
 	},
 });
+
+export function createAuthMiddleware() {
+	return factory.createMiddleware(async (c, next) => {
+		// WebSocket clients pass the token as a query param (no Authorization
+		// header on upgrade requests). Normalise into a Bearer header so
+		// Better Auth's bearer() plugin handles extraction uniformly.
+		const wsToken = c.req.query('token');
+		const headers = wsToken
+			? new Headers({ authorization: `Bearer ${wsToken}` })
+			: c.req.raw.headers;
+
+		const result = await auth.api.getSession({ headers });
+		if (!result) return c.json({ error: 'Unauthorized' }, 401);
+
+		c.set('user', result.user);
+		c.set('session', result.session);
+		await next();
+	});
+}
